pyproject.toml
[project]
name = "pdf-image-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "openai>=2.16.0",
    "pymupdf>=1.26.7",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
